Main API functional, next: organize project in clean structure and develop CLI interface that uses mainAPI class.

+++ mainAPI.py
@@ -62,6 +62,9 @@ class mainAPI():
 
     def deleteTask(self, task):
 
+        if(not self.inDb('Tasks', 'task', task)):
+            return 'Error: the specified entry does not exist in the database'
+        
         try:
             self.conn.execute("delete from Tasks where task='{0}'".format(task))
             return 'Success'
@@ -72,6 +75,9 @@ class mainAPI():
 
     def modifyTask(self, task, mod_task, mod_utgid, mod_command, mod_task_parameters, mod_command_parameters, 
                 mod_description):
+
+        if(not self.inDb('Tasks', 'task', task)):
+            return 'Error: the specified entry does not exist in the database'
         
         try:
             self.conn.execute("update Tasks set task='{0}', utgid='{1}', command='{2}', task_parameters='{3}', \
@@ -86,6 +92,9 @@ class mainAPI():
 
     def getTask(self, task):
 
+        if(not self.inDb('Tasks', 'task', task)):
+            return 'Error: the specified entry does not exist in the database'
+
         try:
             query = self.conn.execute("select * from Tasks where task='{0}'".format(task))
             result = {'data': [dict(zip(tuple(query.keys()), i)) for i in query.cursor]}
@@ -96,6 +105,9 @@ class mainAPI():
     # ---------------------------------------------------------------------------------------------------------------------
 
     def assignTask(self, task, task_set):
+
+        # if(not self.inDb('Tasks', 'task', task)):
+        #     return 'Error: the specified entry does not exist in the database'
 
         try:
             self.conn.execute("insert into Tasks_to_Task_Sets values ('{0}', '{1}')".format(task, task_set))
@@ -108,7 +120,8 @@ class mainAPI():
     def unassignTask(self, task, task_set):
 
         try:
-            self.conn.execute("delete from Tasks_to_Task_Sets where task='{0}' and task_set='{1}'".format(task, task_set))
+            self.conn.execute("delete from Tasks_to_Task_Sets where task='{0}' and task_set='{1}'".format(task, 
+                            task_set))
             return 'Success'
 
         except Exception as e: return e
@@ -129,6 +142,9 @@ class mainAPI():
 
     def deleteSet(self, task_set):
 
+        if(not self.inDb('Task_Sets', 'task_set', task_set)):
+            return 'Error: the specified entry does not exist in the database'
+
         try:
             self.conn.execute("delete from Task_Sets where task_set='{0}'".format(task_set,))
             return 'Success'
@@ -139,9 +155,12 @@ class mainAPI():
 
     def modifySet(self, task_set, mod_task_set, mod_description):
 
+        if(not self.inDb('Task_Sets', 'task_set', task_set)):
+            return 'Error: the specified entry does not exist in the database'
+        
         try:
-            self.conn.execute("modify Task_Sets set task_set='{0}', description='{1}' where task_set='{2}'".format(mod_task_set,
-                            mod_description, task_set))
+            self.conn.execute("modify Task_Sets set task_set='{0}', description='{1}' where task_set=\
+                            '{2}'".format(mod_task_set, mod_description, task_set))
             return 'Success'
 
         except Exception as e: return e
@@ -150,6 +169,9 @@ class mainAPI():
 
     def getSet(self, task_set):
         
+        if(not self.inDb('Task_Sets', 'task_set', task_set)):
+            return 'Error: the specified entry does not exist in the database'
+
         try:
             query = self.conn.execute("select * from Task_Sets where task_set='{0}'".format(task_set,))
             result = {'data': [dict(zip(tuple(query.keys()), i)) for i in query.cursor]}
@@ -160,6 +182,9 @@ class mainAPI():
     # ---------------------------------------------------------------------------------------------------------------------
 
     def assignSet(self, task_set, node_class):
+
+        # if(not self.inDb('Task_Sets', 'task_set', task_set)):
+        #     return 'Error: the specified entry does not exist in the database'
 
         try:
             self.conn.execute("insert into Task_Sets_to_Classes values ('{0}','{1}')".format(task_set, node_class))
@@ -172,8 +197,8 @@ class mainAPI():
     def unassignSet(self, task_set, node_class):
 
         try:
-            self.conn.execute("delete from Task_Sets_to_Classes where task_set='{0}' and class='{1}'".format(task_set, 
-                            node_class))
+            self.conn.execute("delete from Task_Sets_to_Classes where task_set='{0}' and class=\
+                            '{1}'".format(task_set, node_class))
             return 'Success'
 
         except Exception as e: return e
@@ -184,7 +209,7 @@ class mainAPI():
 
     def addClass(self, node_class, description):
 
-        try
+        try:
             self.conn.execute("insert into Classes values ('{0}','{1}')".format(node_class, description))
             return 'Success'
 
@@ -193,6 +218,9 @@ class mainAPI():
     # ---------------------------------------------------------------------------------------------------------------------
 
     def deleteClass(self, node_class):
+
+        if(not self.inDb('Classes', 'class', node_class)):
+            return 'Error: the specified entry does not exist in the database'
 
         try:
             self.conn.execute("delete from Classes where class='{0}'".format(node_class,))
@@ -204,9 +232,12 @@ class mainAPI():
   
     def modifyClass(self, node_class, mod_node_class, mod_description):
 
+        if(not self.inDb('Classes', 'class', node_class)):
+            return 'Error: the specified entry does not exist in the database'
+
         try:
-            self.conn.execute("update Classes set class='{0}', description='{1}' where class='{2}'".format(mod_node_class,
-                            mod_description, node_class))
+            self.conn.execute("update Classes set class='{0}', description='{1}' where class=\
+                            '{2}'".format(mod_node_class, mod_description, node_class))
             return 'Success'
 
         except Exception as e: return e
@@ -214,6 +245,9 @@ class mainAPI():
     # ---------------------------------------------------------------------------------------------------------------------
     
     def getClass(self, node_class):
+
+        if(not self.inDb('Classes', 'class', node_class)):
+            return 'Error: the specified entry does not exist in the database'
 
         try:
             query = self.conn.execute("select * from Classes where class='{0}'".format(node_class,))
@@ -225,6 +259,9 @@ class mainAPI():
     # ---------------------------------------------------------------------------------------------------------------------
     
     def assignClass(self, node_class, node_regex):
+
+        # if(not self.inDb('Classes', 'class', node_class)):
+        #     return 'Error: the specified entry does not exist in the database'
 
         try:
             self.conn.execute("insert into Classes_to_Nodes values ('{0}', '{1}')".format(node_class, node_regex))
@@ -286,10 +323,19 @@ class mainAPI():
             return self.json.dumps(result)
 
         except Exception as e: return e
-            
+
     # =====================================================================================================================
     #                          Helper Methods
     # =====================================================================================================================
 
+    def inDb(self, table, column, value):
 
+        arr = []
+        result = self.conn.execute('SELECT ' + column + ' FROM ' + table + ' WHERE ' + column + '="' + value + '\"')
+        for row in result:
+            arr.append(row)
+        if len(arr) == 0:
+            return False
+        else:
+            return True
